+++ src/helpers.js
@@ -1,0 +1,12 @@
+import { spawn } from 'child_process'
+
+export function cmd(cmd, args = [], options = {}) {
+  return new Promise((resolve, reject) => {
+    const child = spawn(cmd, args, options)
+    let data = ''
+
+    child.stdout.on('data', buffer => data += buffer.toString())
+    child.stdout.on('end', () => resolve(data))
+    child.on('error', reject)
+  })
+}
