+++ src/index.js
@@ -1,21 +1,16 @@
-import { spawn } from 'child_process'
 import { parseCommits, LOG_FORMAT } from './commits'
+import { cmd } from './helpers'
 
-
-function cmd (cmd, args = [], options = {}) {
-  return new Promise((resolve, reject) => {
-    const child = spawn(cmd, args, options)
-    let data = ''
-
-    child.stdout.on('data', buffer => data += buffer.toString())
-    child.stdout.on('end', () => resolve(data))
-    child.on('error', reject)
-  })
-}
-
-function getCommits () {
+function getCommits() {
   return cmd('git', ['log', '--shortstat', '--pretty=format:' + LOG_FORMAT]).then(parseCommits)
 }
+
+function getOrigin() {
+   return cmd('git', ['remote', 'show', 'origin']).then(remote => {
+     const url = 'https://' + remote.match(/github.com(:|\/)[^\/]+\/[^\.]+/)[0].replace(':', '/');
+     return url
+   })
+ }
 
 Promise.all([ getCommits()]).then(([ commits]) => {
   console.log(commits)
